Move the `data-model-client` module under the `@bayou` namespace.
