+++ local-modules/ui-components/Owner/index.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 
-import { OwnerState } from 'data-model-client';
+import { OwnerState } from '@bayou/data-model-client';
 import styles from './owner.module.less';
 
 /**
